accel/amdxdna: move AIE tile read/write to amdxdna_tile_read_write.c
Extract amdxdna_aie_tile_read() and amdxdna_aie_tile_write() along
with their supporting helpers and walk struct from aie.c into a
dedicated amdxdna_tile_read_write.c / amdxdna_tile_read_write.h pair.
Update aie2_pci.c and aie4_pci.c to include the new header, and add
the new object to Kbuild.
