+++ drivers/accel/amdxdna/Kbuild
@@ -55,6 +55,7 @@ amdxdna-$(OFT_CONFIG_AMDXDNA_PCI) += \
 	amdxdna_dpt.o \
 	amdxdna_error.o \
 	amdxdna_coredump.o \
+	amdxdna_tile_read_write.o \
 	npu1_regs.o \
 	npu3_regs.o \
 	npu4_regs.o \
